Create README.md
This uses the SSPI Schannel C/C++ libraries which connects to the servers on port 443 using RSA SSL/TLS security contexts. It will fetch images from google and dump the binary image to a specified file. You can play around with it and make it work how you want it to. It works with no errors on C++. You can comment on this respiratory if you need help with something.